kuran_kursu_app/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0

# Database
pymysql==1.1.0
cryptography==41.0.7

# Background Tasks
celery==5.3.4
redis==5.0.1

# Forms & Validation
WTForms==3.1.1
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

# Excel & Reports
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date & Time
python-dateutil==2.8.2

# Security
werkzeug==3.0.1
bcrypt==4.1.2
